Revision indexing is now based on information on the revision table (although command based indexing is still usable), removing revisions from index is still command based. Commands don't hang around in database anymore after they have been completed. Revision indexing is now threadable and by default three separate indexers are used for revision based indexing. Variable files can not be changed to non variable files anymore. Variable file language can not be changed anymore (same file can not be responsible for multiple variables revisionables).

+++ metka/src/main/java/fi/uta/fsd/metka/transfer/settings/OpenIndexCommandsResponse.java
@@ -31,14 +31,14 @@ package fi.uta.fsd.metka.transfer.settings;
 import fi.uta.fsd.metka.storage.repository.enums.ReturnResult;
 
 public class OpenIndexCommandsResponse {
-    private Integer openCommands;
+    private Long openCommands;
     private ReturnResult result;
 
-    public Integer getOpenCommands() {
+    public Long getOpenCommands() {
         return openCommands;
     }
 
-    public void setOpenCommands(Integer openCommands) {
+    public void setOpenCommands(Long openCommands) {
         this.openCommands = openCommands;
     }
